chore: refactor to common component comp structure
Aligns the FileUploader components with folder/file name
structure of other connected components by:
* Moving all components into their own ComponentName.tsx files
* Move FileUploader and UploadMessage into separate component folders
