+++ packages/react/src/components/Storage/FileUploader/FileUploader/index.ts
@@ -1,0 +1,1 @@
+export { FileUploader } from './FileUploader';
