+++ packages/react/src/components/Storage/FileUploader/UploadDropZone/index.ts
@@ -1,0 +1,1 @@
+export { UploadDropZone } from './UploadDropZone';
