+++ packages/react/src/components/Storage/FileUploader/UploadMessage/index.ts
@@ -1,0 +1,1 @@
+export { UploadMessage } from './UploadMessage';
