+++ packages/react/src/components/Storage/FileUploader/UploadPreviewer/index.ts
@@ -1,0 +1,1 @@
+export { UploadPreviewer } from './UploadPreviewer';
