+++ packages/react/src/components/Storage/FileUploader/UploadTracker/index.ts
@@ -1,0 +1,1 @@
+export { UploadTracker } from './UploadTracker';
